feat: Use last_check_in for staleness and adapt code to use feat flag
