+++ app/queue/events.py
@@ -46,6 +46,7 @@ class SerializedHostSchema(Schema):
     provider_type = fields.Str()
     created = fields.Str()
     updated = fields.Str()
+    last_check_in = fields.Str()
     stale_timestamp = fields.Str()
     stale_warning_timestamp = fields.Str()
     culled_timestamp = fields.Str()
